database change to local

+++ db/migrate/20141007160739_change_events.rb
@@ -1,12 +1,18 @@
 class ChangeEvents < ActiveRecord::Migration
-  def up    
-    remove_column :events, :rating_phase, :string    
+  def up   
+    if column_exists? :events, :rating_phase 
+      remove_column :events, :rating_phase, :string    
+    end
     add_column   :events, :rating_phase, :integer, default: 0 
     
-    remove_column :events, :rating_system, :string    
+    if column_exists? :events, :rating_system 
+      remove_column :events, :rating_system, :string    
+    end
     add_column :events, :rating_system, :integer
     
-    remove_column :events, :voting_range, :string    
+    if column_exists? :events, :voting_range 
+     remove_column :events, :voting_range, :string   
+    end 
     add_column :events, :voting_range, :integer
     
   end
